Move wait cli command to task wait
