+++ cmd/root.go
@@ -8,7 +8,6 @@ import (
 	"github.com/ohsu-comp-bio/funnel/cmd/server"
 	"github.com/ohsu-comp-bio/funnel/cmd/task"
 	"github.com/ohsu-comp-bio/funnel/cmd/termdash"
-	"github.com/ohsu-comp-bio/funnel/cmd/wait"
 	"github.com/spf13/cobra"
 )
 
@@ -29,5 +28,4 @@ func init() {
 	RootCmd.AddCommand(server.Cmd)
 	RootCmd.AddCommand(task.Cmd)
 	RootCmd.AddCommand(termdash.Cmd)
-	RootCmd.AddCommand(wait.Cmd)
 }
