+++ cmd/task/task.go
@@ -21,4 +21,5 @@ func init() {
 	Cmd.AddCommand(createCmd)
 	Cmd.AddCommand(getCmd)
 	Cmd.AddCommand(cancelCmd)
+	Cmd.AddCommand(waitCmd)
 }
